+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/SchemaProperties.scala
@@ -35,6 +35,14 @@ final case class SchemaProperties[T](
     properties(tag).asInstanceOf[S]
   }
 
+  def getOrNone[S <: SchemaProperty[T, S]](implicit tag: ClassTag[S]): Option[S] = {
+    if (properties.contains(tag)) {
+      Some(properties(tag).asInstanceOf[S])
+    } else {
+      None
+    }
+  }
+
   def has[S <: SchemaProperty[T, S]](implicit tag: ClassTag[S]): Boolean = {
     properties.contains(tag)
   }