+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New `getOrNone` method for `SchemaProperties` which returns an `Option`
+
 ### Changed
 - Moved ``BuildInfo`` class to inside `discovery` package
 - Update assembly JAR configuration to change name